+++ Project/Game/Objects/Tornado/Tornado.h
@@ -1,38 +1,31 @@
 #pragma once
 
-#include "GameObjectManager.h"
-
-#include "Camera/GameCamera.h"
-
+#include "GameObjectManager/BaseGameObject.h"
 
 /// ===================================================
 /// 前方宣言
 /// ===================================================
-
 class Player;
-class DirectionalLight;
 
 
 /// ===================================================
-/// ゲームカメラのアクションを行うクラス
+/// 竜巻を表現するクラス
 /// ===================================================
-class GameCameraState final : public BaseGameObject {
+class Tornado final : public BaseGameObject {
 public:
 	/// ===================================================
 	/// public : methods
 	/// ===================================================
 
-	GameCameraState() { CreateTag(this); }
-	~GameCameraState() {}
+	Tornado() { CreateTag(this); }
+	~Tornado() {}
 
 	void Initialize() override;
 	void Update()     override;
-	void LastUpdate() override;
 	void Debug()      override;
 
-	void SetGameCamera(GameCamera* _gameCamera);
+
 	void SetPlayer(Player* _player);
-	void SetDirectionalLight(DirectionalLight* _directionalLight);
 
 private:
 	/// ===================================================
@@ -40,8 +33,19 @@ private:
 	/// ===================================================
 
 	/// other class pointer
-	Player*           pPlayer_     = nullptr;
-	GameCamera*       pGameCamera_ = nullptr;
-	DirectionalLight* pLight_      = nullptr;
+	Player* pPlayer_ = nullptr;
+
+	/// tarnsform setting
+	Quaternion quaternionLocalX_ = { 0.0f, 0.0f, 0.0f, 1.0f };
+	Quaternion quaternionLocalY_ = { 0.0f, 0.0f, 0.0f, 1.0f };
+
+	float localYAngle_ = 0.0f;
+
+	/// action param setting
+	float scaleScaler_ = 1.0f;
+	float eacSpeed_; /// "eac == expansion and contraction"
+
+	float maxScale_;
+	float minScale_;
 
 };
